+++ src/util/sexp/index.ts
@@ -235,7 +235,7 @@ export class SExpUtil {
     }
 
     // https://github.com/Chia-Network/chia-blockchain/blob/main/chia/wallet/puzzles/p2_delegated_puzzle_or_hidden_puzzle.clvm.hex
-    public readonly P2_DELEGATED_PUZZLE_OR_HIDDEN_PUZZLE_PROGRAM = this.fromHex("ff02ffff01ff02ffff03ff0bffff01ff02ffff03ffff09ff05ffff1dff0bffff1effff0bff0bffff02ff06ffff04ff02ffff04ff17ff8080808080808080ffff01ff02ff17ff2f80ffff01ff088080ff0180ffff01ff04ffff04ff04ffff04ff05ffff04ffff02ff06ffff04ff02ffff04ff17ff80808080ff80808080ffff02ff17ff2f808080ff0180ffff04ffff01ff32ff02ffff03ffff07ff0580ffff01ff0bffff0102ffff02ff06ffff04ff02ffff04ff09ff80808080ffff02ff06ffff04ff02ffff04ff0dff8080808080ffff01ff0bffff0101ff058080ff0180ff018080");
+    public readonly P2_DELEGATED_PUZZLE_OR_HIDDEN_PUZZLE_PROGRAM_MOD = this.fromHex("ff02ffff01ff02ffff03ff0bffff01ff02ffff03ffff09ff05ffff1dff0bffff1effff0bff0bffff02ff06ffff04ff02ffff04ff17ff8080808080808080ffff01ff02ff17ff2f80ffff01ff088080ff0180ffff01ff04ffff04ff04ffff04ff05ffff04ffff02ff06ffff04ff02ffff04ff17ff80808080ff80808080ffff02ff17ff2f808080ff0180ffff04ffff01ff32ff02ffff03ffff07ff0580ffff01ff0bffff0102ffff02ff06ffff04ff02ffff04ff09ff80808080ffff02ff06ffff04ff02ffff04ff0dff8080808080ffff01ff0bffff0101ff058080ff0180ff018080");
     // https://github.com/Chia-Network/chia-blockchain/blob/5f4e39480e2312dc93a7b3609bcea576a9a758f9/chia/wallet/puzzles/p2_delegated_puzzle_or_hidden_puzzle.py#L70
     public readonly DEFAULT_HIDDEN_PUZZLE_PROGRAM = this.fromHex("ff0980");
     public readonly DEFAULT_HIDDEN_PUZZLE_HASH = "711d6c4e32c92e53179b199484cf8c897542bc57f2b22582799f9d657eec4699";
@@ -266,16 +266,23 @@ export class SExpUtil {
             key : this.calculateSyntheticPublicKey(key, this.DEFAULT_HIDDEN_PUZZLE_HASH);
 
         return this.curry(
-            this.P2_DELEGATED_PUZZLE_OR_HIDDEN_PUZZLE_PROGRAM,
+            this.P2_DELEGATED_PUZZLE_OR_HIDDEN_PUZZLE_PROGRAM_MOD,
             [
                 SExp.to(Bytes.from(Util.key.publicKeyToHex(syntheticPublicKey), "hex")),
             ]
         );
     }
+    // https://github.com/Chia-Network/chia-blockchain/blob/main/chia/wallet/puzzles/p2_conditions.clvm.hex
+    public readonly P2_CONDITIONS_PROGRAM = this.fromHex("ff04ffff0101ff0280");
     public standardCoinSolution(conditions: SExp[]): SExp {
         return SExp.to([
             SExp.to([]),
-            SExp.to([1, ...conditions]), // 1 = q (this is a puzzle, not a solution)
+            this.run(
+                this.P2_CONDITIONS_PROGRAM,
+                SExp.to([
+                    SExp.to(conditions),
+                ])
+            ),
             SExp.to([])
         ]);
     }